+++ proxy/build.rs
@@ -1,0 +1,10 @@
+fn main() -> Result<(), Box<dyn std::error::Error>> {
+    #[cfg(feature = "grpc")]
+    {
+        tonic_build::configure()
+            .build_server(true)
+            .build_client(true)
+            .compile_protos(&["proto/control_plane.proto"], &["proto"])?;
+    }
+    Ok(())
+}